feature(log) Add memory logging for feed/rss request
